Add: REST API Framework Use Volley
1. at retailer side, add post-dish buttonn and corresponding activity
2. at retailer side, add delete post callback
3. add ratingbar listener for customer
4. fix one bug for using Toast in doInBackground for Customer Map tab
5. enable twitter login via retailer and customer
6. add user ID for twitter login
7. Add: REST API Framework Use Volley
8. Add: Login Create User Finished
9. Add: Fetch Dish Information from Backend Server
10. Add: Post Rating
11. Add: Payment by Paypal Integration

+++ EasyOrder_Android/app/src/main/java/edu/cmu/EasyOrder_Android/RetailerDishListAdapter.java
@@ -49,7 +49,7 @@ public class RetailerDishListAdapter extends ArrayAdapter<Dish> {
 
         RatingBar ratingBar = (RatingBar) v.findViewById(R.id.retailer_ratingBar);
         ratingBar.setIsIndicator(true);
-        ratingBar.setRating(dish.getRate());
+        ratingBar.setRating((float)dish.getRate());
 
         return v;
     }
